requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
axial_positional_embedding==0.2.1
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmdstanpy==1.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.2
datasets==4.4.1
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
einops==0.4.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
holidays==0.86
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
llvmlite==0.46.0
local-attention==1.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.0
notebook_shim==0.2.4
numba==0.63.0
numpy==1.23.5
nvidia-ml-py==13.590.44
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.5
patool==1.12
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.1
pmdarima==2.1.1
prettytable==3.17.0
product_key_memory==0.1.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.2.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyecharts==2.0.9
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
reformer-pytorch==1.4.4
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
simplejson==3.20.2
six==1.17.0
sktime==0.16.1
soupsieve==2.8
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.6
swanlab==0.7.5
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.3.0
#torch==2.9.1+cu130
#torchvision==0.24.1+cu130
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
